First iteration of dashboard (mock) (#15)
Replaces the old Home screen with a prototype of the new dashboard/home
page with very basic styling and almost no functionality. Adds several
widgets and styling decisions that can be reused in other parts of the
application in order to maintain consistency. Specially important are
the constants introduced, which encapsulate themes, styles, dimensions,
and other design choices.