Add Firebase Core and Crashlytics integration
Note that the google-services.json (Android) and GoogleService-Info.plist (iOS) are not included here because these are to be generated using the Firebase CLI to retrieve them from the project hosted in Firebase. These can be added to version control, but they expose our API Key and is not ideal for an Open Source project.