Refactor GraphQL user queries and data models (#36)
The purpose of this PR is to introduce some useful OOP design patterns,
strong typing, and structure into the existing GraphQL query operations
for _User_. The remaining models and operations will be later refactored
in the same manner if this PR is accepted.

The big motivation behind this PR is the gotcha of using GraphQL, which
is that queries can return any number of fields, and so a single model
class (e.g. UserModel) cannot accurately represent the response received
from the server. Some queries will populate some fields of the
UserModel, some will populate others, but very rarely you will have an
instance with all of the fields available. Meaning that all fields of
the model must be nullable, and that caller must be extremely careful to
access only the fields that contain data.

Having each query return a very specific result object with only the
fields that they need removes this problem. It also makes the providers
easier to use since the caller only has access to what they need,
instead of having to manage a JSON object and extracting fields
directly.

### Key changes

- **BaseProvider** abstract class that is extended by other provider
classes. It contains a runQuery function that takes a BaseOperation as
input, and returns a BaseResult as output.
- **BaseOperation** abstract class that represent a single GraphQL
query. It contains the logic to convert a GQL json response into the
corresponding BaseResult object.
- **BaseResult** abstract class that represents the response of a GQL
query. Contains only the exact fields that the query can return, with
proper typing and nullable fields.
- **Automatic query retries** while remaining on the 'Loading' state.
Once the maximum number of attempts is reached the state changes to
'Error'.
- **Refetch** function callback is exposed to the caller to facilitate
manually retrying (or refreshing) from the onData callback, without
triggering the operation all over again.
- **FetchMore** function callback is exposed to callers to facilitate
pagination of data. This might be revisited later to add more
abstraction.

### Next steps (if this is approved)

- Implement a similar structure for Mutations
- Implement for all data models and operations
- Explore ways to make better use of the ChangeNotifiers, perhaps by
breaking them up into smaller pieces and leveraging the notify function