Use GraphQL Codegen to create model classes for User queries/mutations (#40)
This introduces the use of the [GraphQL
Codegen](https://pub.dev/documentation/graphql_codegen/latest/) package
to generate model classes for queries/mutations, as defined in the
schema.graphql and operations.graphql files.

This build upon the previous iteration which introduced the use of
unique model classes for each GraphQL operation. By completely
automating the creation of model classes it will now take a few lines of
code to run a GraphQL operation, and receive its result in a properly
typed object.

### Key Changes

- Should eliminate the need to manually create model classes for GraphQL
operations.
- We can easily have multiple versions of the same query (with different
fields returned), since the codegen will take care of creating the model
classes for each version with their corresponding fields.
- Will eventually make everything under the data/models directory
obsolete, except for the provider classes. At that point these might be
moved to a different directory structure.
- All operations now return a OperationResult<T> object, which contains
the model class that represents the output of the operation.
- BaseOperation and BaseResult (from previous iteration) are now fully
replaced with generated classes. The main logic for executing GraphQL
operations remains mostly unchanged.