UI for mentor/entrepreneur profile quick view card
The widget parameters determine if the card displays the entrepreneur version or the mentor version. Note that the previous explore page was completely replaced with three hardcoded cards, just as a demonstration of how they would look. Once the Explore page frame is completed please replace the build Widget at explore.dart.