UI for mentor/entrepreneur profile quick view card (#51)
The widget parameters determine if the card displays the entrepreneur
version or the mentor version. Note that the previous explore page was
completely replaced with three hardcoded cards, just as a demonstration
of how they would look. Once the Explore page frame is completed please
replace the build Widget at explore.dart.

Some misc. additions in this PR:

1. Added a blank-avatar.png asset that we can use as the default
whenever the user doesn't have an avatar. The actual image can be
replaced later, but we should be already implementing the null checking
and defaulting to this asset.
2. Added an enums.dart to the constants library. Right now the only enum
is UserType (mentor or entrepreneur). We will be doing a lot of checks
for user type so it made sense to create this enum. Feel free to add
more as needed.


![image](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/646f1d95-d62a-4c2d-ac4a-4d1994b2bf1d)