Login Page UI changes (#48)
This is the new UI for the log in page for the Flutter app.
<img width="500" alt="Screenshot 2023-07-06 at 8 06 27 PM"
src="https://github.com/micromentor-team/mm-flutter-app/assets/63002494/8c50f86d-3888-4344-93c1-e470965164dc">


I need feedback on the following:
1) I used an external package for the social icons called
"FlutterSocialButton": is that recommended? If not, I can use some
external image icons instead.
2) How do you localize text fields within the form, for instance: "Email
or phone number" or "password"? The standard l10n.field_name is giving
an error, namely "201:52: Error: Not a constant expression."

---------

Co-authored-by: Rupal Gupta <rupalgupta99@gmail.com>