Major updates to navigation and state management for dynamic Scaffold (#63)
This PR introduces many changes to facilitate maintaining the Scaffold
state when pushing/popping pages to the navigation stack using GoRouter.
The need for this arises from the fact that the Scaffold components
change across pages: some have an AppBar with a title (Inbox), some have
a TabBar at the bottom of the AppBar (Invites Received/Sent), and some
introduce a custom AppBar that displays elements based on the route
(ChannelMessages).

The most notable changes introduced are the following:

- ScaffoldModel (ChangeNotifier) class was introduced to notify the main
Scaffold when changes are needed to the AppBar or Drawer.
- Factory methods were added to utilities to facilitate creating
repetitive AppBar and Drawer components.
- RouteAwareMixin was introduced to make all Screen widgets observe the
router and update the Scaffold whenever a push/pop occurs. Some
boilerplate code had to be added to all Screen widgets.
- The first parameterized route was introduced (ChannelMessages), which
allows navigating to a messages screen for a Channel with the ID given
in the route path.
- Created MaterialPage widgets for all routes, and added placeholder
Screen widgets for Explore, Profile, and Progress. Removed the obsolete
Profile widgets created in the first iteration.