Add filters to the explore page (#66)
This PR contains implementation for "simple" search filters (once it's
approved I'll make a quick PR adding the "advanced" view). It's a draft
for now as there are a few things missing:
- Internationalization
- Adjusting colors
- Proper behavior of the "All" button
- Some code cleanup

Otherwise it's fully functional and works with the main Explore view
using a `Provider`. One thing missing is a dropdown button for
autocomplete but I found working on that highly overwhelming. The
autocomplete will trigger with empty input nonetheless, so I think it's
redundant.

This PR is a draft, so please don't review it (you are free to test if
you are curious). I'll finish it tomorrow.

---------

Co-authored-by: Miguel Rosario <miguel.a.rosario.aponte@gmail.com>