Create mock resolvers for replying, deleting, and updating messages (#76)
This allows deleting/undeleting, editing messages, creating replies, and
marking messages as seen in the mock server. These changes persist as
long as the server is running.

I need confirmation on the logic for the backend. I know these fields
are all that is needed for the frontend, but I would like to know if
there are side effects to running these operations that we need to be
aware of. For example, does editing a message update the `editedAt` and
`updatedAt` fields? Does deleting or undeleting a message update the
`deletedAt` and `deletedBy` fields?

Also, is the backend able to tell the difference between a `deletedAt`
field that is null because it wasn't set by the caller in the
`updateChannelMessage` mutation, or because it was specifically set to
null by the caller (to undelete the message)?