Miscellaneous fixes in the home page (#87)
Found a few bugs when testing with the real backend:

- The queryAllUsers call doesn't work without a 'filter' parameter.
Added an inconsequential parameter just so that it won't crash. Though
we should definitively revise this later to include pagination.
- AvatarUrl in the Recommended Mentor tile was defined as a required
parameter, but in practice it can be null and that causes an exception.
Made it a nullable field.
- Added a default empty avatar image for header, recommended mentors,
and upcoming sessions.