Archiving/unarchiving of Inbox chats (#98)
This PR implements functionality to archive/unarchive a channel either
from the list of channels (with a swipe gesture) or from the channel
messages screen (using three dots menu). For now all that
archiving/unarchiving does is move the channel in or out of the Archived
list.

The AppBar and Drawer are also updated to show channel notification
count. The number corresponds to the number of unread messages from
unarchived channels only.

The mock server was updated to allow testing these implementations.


[chat-archive.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/c7b7e19e-ea25-4d31-b18f-37fae20c3e4e)