Improved event handling for channel messages (#102)
Introduces channelChanged subscription event filtering so that models
(a.k.a. ChangeNotifiers for UI components) only process events that are
relevant to them. Since the type of event and channel/message id is now
included in the subscription, event handling is now more specific and
efficient: we now refresh only the object that changed instead of
refreshing all of the data for every event received.

Note that the frontend uses the models to implement optimistic updates,
so that the UI is updated without querying the backend. An eventual
'refresh' with the backend data ensures that both backend and frontend
are synced.

Includes several other fixes and improvements, but as indicated in the
TODOs, there is still more work to do in regards to optimization.