Backend integration for Inbox received invites (#104)
Fetches the pendingInvitations list from myInbox and displays it in the
Invites / Received tab. The query is performed once per page load, it is
not yet updated in real-time like the chat given that there is little
gain relative to the the cost of listening for channel events to refresh
this particular page. This can be added later if needed.

The notification count in the 'hamburger' icon and the drawer are also
updated to reflect the number of notifications from pending invites
(Hamburger notifications = all notifications minus the ones from the
current page, which is why it changes depending on the page)


[invites_received.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/5054a5b1-0f4b-44ff-ae31-841be96263a7)