Update schema and reconcile breaking changes (#112)
Updates the schema to the latest version. The following breaking changes
were also addressed:

- ~~The User type no longer contains the **companies** field, so an
additional query had to be made to build received invitation details.~~
Schema updated again so this is no longer needed.
- DateTimeISO type had to be specified in the codegen builder, and in
the mock server.

EDIT: After the reverts and revisions the net effect of this PR is an
updated schema, addition of DateTimeISO, and some minor formatting
fixes. 🤷

---------

Co-authored-by: Chris <andersan@umich.edu>