NavigationMixin to manage PageRoutes more efficiently (#130)
PageRoutes that are not in top of the navigation stack will now have
their `build` and `didDependenciesChange` methods blocked, which
improves efficiency dramatically. One serious issue we were having is
that every single page added to the navigation stack continued
rebuilding and fetching data (`didChangeDependencies` keeps getting
called, that's just how Flutter works with stateful widgets), which made
the performance slow down to almost a halt as the user pushed more and
more pages onto the stack.

Addionally, the RouteAwareMixin (now called **NavigationMixin**) was
updated so that PageRoutes don't have to manually refresh the scaffold
after router pops and pushes.

Some gains from this change:
- Much cleaner and streamlined code.
- The Scaffold and the PageRoute widget are build at the same time,
which should eliminate syncing issues that arose from having the
Scaffold and screen build separately.
- Blocking background rebuilds and data refreshes will provide a MASSIVE
performance improvement. In fact, this might resolve (or at least
alleviate significantly) the latency issues experienced in the last
testing session.

### A few more details on how the stateful widget optimization works...

Stateful widgets maintain their state and rebuild constantly even if on
the background. Particularly, the `didDependenciesChange` and the
`build` methods get called on each new push/pop that occurs, even for
other pages. By setting the `maintainState` property of the PageRoute to
false, we eliminate these calls whenever the page is not visible.
However, there is still at least one additional rebuild that occurs when
the page is being removed from the top of the stack, meaning that
everytime a new push/pop occurs, the two pages involved have their build
methods called almost simultaneously, which is why sometimes you ended
up with the header of the previous page on the new page. The
NavigationMixin `pageRoute.isCurrent` is used to block these additional
frames of the now obsolete page and further improve performance. The net
effect is that there is only one call to `didDependenciesChange` per
page opened.

** Note that the NavigationMixin can only mix with the "screen" widgets
under the AppWrapper (all PageRoutes defined inside the ShellRoute),
which should be enough since these are the parent widgets of all other
components.