Message context menu, swipe-to-reveal tiles, and other Inbox goodies (#140)
This includes miscellaneous improvements to the Inbox, some of which
were requested by UX from the testing session:

- User can create a new invitation after withdrawing a previous
invitation.
- The swipe-left-to-archive gesture was replaced with a swipe-to-reveal
action, similar to how iOS handles this type of gesture. This was
requested to have the user take the extra step of clicking on the icon
to confirm the action.
- Archiving/unarchiving a chat from the overflow menu takes the user to
the Inbox chat list.
- Swipe gestures on messages are removed, and replaced with a context
menu that is activated by long press. Note the following specifics of
the context menu:

1. Undeleted messages sent by authenticated user show Edit and Delete
options.
2. Deleted messages sent by authenticated user don't show any options
(undelete is not supported).
3. Messages sent by other user don't show any options for now. Later we
will add options for emoji reactions.
4. The Reply functionality is no longer supported. The code for it is
still there, but UX asked to remove the ability to reply while they
figure out the correct flow for it. This option can be easily added to
the context menu later if needed.

- Finally, there was a lot of cleanup and code simplification.
MessagePeeker, MessageHoverover, and MessageDetails are now deleted and
replaced by the MessageContextMenu

This video demonstrates the message context menu:


[chat_message_context.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/9dc742b7-b235-48dd-804a-c93739b9e9cd)

This video demonstrates the swipe-to-reveal tiles:


[swipe-to-reveal-tiles.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/d1a8d999-81f1-44aa-a0f6-3733ba8895d0)

---------

Co-authored-by: Rupal Gupta <63002494+rupalg@users.noreply.github.com>