Inbox navBar notification bubbles, more efficient inbox model, UI improvements (#147)
Adds several improvements to the Inbox:

- Inbox notification bubble now appears in the navigation bar at the
bottom of the screen.
- Notification bubbles for hamburger icon and navBar now show total
number of unread chat messages + received invitations. This will be
modified later to only include _**unread**_ received invitations.
- Following the latest UX design, edited and deleted messages now appear
differently
- Added background to emoji-only messages. Having transparent background
was not specified by UX, and made it more difficult to trigger the
hard-press gesture.
- Added an InboxModel to query channels and invitations more
efficiently, since it is now done during application start and during
relevant events.
- Introduced the use of **Selectors** to replace Consumer widgets in
parts of the app that only depend on one or two fields of a model. The
reduced number of rebuilds help in further improving performance.

The following video demonstrates visual changes:


[inbox-notifications.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/37bfc6f4-4afe-41d3-84c1-8b4bf8fb78ca)

---------

Co-authored-by: Rupal Gupta <63002494+rupalg@users.noreply.github.com>