Signup flow integration with backend (entrepreneur & mentor) (#158)
Allows navigating through the signup flow to create and update a user
with group specific fields.

The **UserRegistrationModel** is used to store the data from each page
of the signup flow, which is then used to create a user and then update
it with additional fields.

There are 3 key points in the signup flow:

1. The **Submit** button in the email verification page triggers the
creation of the user in the backend with just the email and the
password. Once user creation succeeds the user navigates to the next
page. If the user closes the app at this point and logs in again, the
app will detect that the signup is not completed and will redirect the
user to the page immediately after the verification.
2. The **We just need a little more information** page (after selecting
Entrepreneur/Mentor) provides the user with the option to skip the rest
of the signup flow. If the user chooses to skip, an updateUser mutation
is triggered with the fields that were collected up to that point, and
once the mutation completes the user is taken to the Explore page. The
user can sign up and access the app normally at this point, but will
need to edit their profile through the Profile page if they wish to add
the data that they skipped over.
3. At the very end of the signup flow the updateUser mutation must be
completed to continue. Just like in the previous case, the user is taken
to the Explore page and is free to use the app normally. Obviously,
their profile will now show additional fields.

UPDATE: Both mentor and entrepreneur flows are working!

Entrepreneur signup demo:

[signup-v2-entrepreneur.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/7255f5a9-93ce-450d-9b7a-8c9928fa51ed)

Mentor signup demo:

[signup-v2-mentor.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/19cc88c0-109d-46f6-b9b0-51e63d13e8d2)