Home page revision and backend integration (#169)
Revises UI paddings and positioning for home page components. Also
integrates the recommended users section with the backend to perform a
search constrained to show mentors to entrepreneurs and vice-versa. Did
some work on the mock server to support latest queries too.

Note that the recommended section will need more work in the future in
order to apply the matching engine and consider other factors such as
existing relationships between users.


[home-revision.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/c0134057-ebd2-4a5d-b122-0bf8bd156912)