Enable marking received invitations as read (#171)
Received invitations are marked as read whenever the user opens the
invitation detail page. At this point the inbox invite notification for
this specific invitation is removed, regardless if the invite was not
accepted or declined. I also added a plain notification bubble in the
inbox received invitations list tiles whenever the invitation hasn't
been opened.

Modified the mock server to support new field readByRecipientAt.

The following video shows the feature in action:


[received-invites-read-receipt.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/a6daa63e-3533-4d18-b2e7-42928442a5b7)