Signup UX revision (1st pass) (#179)
This takes care of the majority of the changes requested by UX for the
signup flow. The flow is shortened significantly, and most screens were
changed.

Key changes included here:

- Preferred language only allows selecting a single value (just like the
backend). There is no option for selecting other spoken languages during
signup.
- There is no email verification.
- It is possible to navigate back to the first page to change the email
even after reaching the last page of the flow. For this reason we now
only create the user at the very end of the flow.
- Forced navigation buttons to remain visible even when keyboard is
onscreen.
- General UI and cleanup improvements

Although it is completely functional, note that this is just a first
pass and there is still many many details left to address: signup method
screen, collapse expertise selection chips, add field validation and
corresponding error messages, scroll entire page up when keyboard pops
up, etc. This will be done in a follow up PR.

The following video illustrates the updated flow for an
**Entrepreneur**:


[signup_rev_1.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/00ee955a-984e-4c0e-8753-d3a381618084)