Signup UX Revision (2nd Pass) (#182)
Further refines the signup flow according to the latest UX revision.

Key changes included here:

- Updated the Social Media SSO buttons in the signin and signup pages
- Added 'dots indicators' to the welcome page, which will be used in the
future to change the slides showing a different picture and text. Right
now these are clickable and change the index accordingly, but the slide
doesn't change yet because we do not have the content for these other
two slides.
- Update the Multi-select chips for the expertises. New version is much
more compact and shows a checkbox for accessibility.
- Included the flutter_markdown package to allow including hyperlinks
and rich formatting as part of the localization strings. This is
important because different languages have different sentence
structures, therefore formatting that is applied only to certain words
needs to be included in the translations as well.
- Allowed pages in the signup flow to be scrolled vertically when there
is not enough space in the screen (e.g. the keyboard is out). The
navigation buttons remain on top always, and there is a subtle fade at
the top of the screen for visual effect when scrolling.

Still pending: 

- field validation with corresponding error messages
- location search

The following video illustrates the entire flow with the changes above:

[signup-rev-2.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/a6aad5b9-10a6-439f-84e1-6674dcca18d8)