Edit Profile pages with backend integration (#184)
This adds all remaining pages to edit profile fields. Profile is updated
in the backend when the user navigates out of each individual edit page.

Note the following:

- Fields that are not working due to being missing in the backend:
Business location, Mentoring preferences, additional expertises, photo
upload
- As of the time this PR was created, there is a bug in the backend that
clears all array fields in the mentor/mentee group memberships whenever
the update mutation is called without including all optional parameters.
This is a bug in the backend, so it is expected that this will be fixed
without needing to update the frontend.
- I did some cleanup on the profile view to remove sections that were no
longer needed (user prompts, expectations), and that were hardcoded
(mentoring preferences, affiliations, business location). This is to
facilitate testing and prevent confusion.
- City lookup is not yet implemented. As of the moment we are allowing
the user to enter any value and it will be stored in the city field.
Once lookup is implemented we will add the region and country fields to
the mix.
- There is no field validation, and there is a lot more polishing to be
done! Consider this a minimum product for the testing session.

The following video demonstrates how a minimal profile (created using
the revised signup flow) can be edited and viewed:


[edit-profile.webm](https://github.com/micromentor-team/mm-flutter-app/assets/25093428/be7ed0a7-b155-46a3-9e02-7ad8c207404e)