Fix late variable error, relative import (#187)
Minor changes - in the debugger there was an exception when
`_tokenSubscriptionStream.cancel();` was called, which this fixes. Not
sure if that has any effect on the app.

In my local, I am unable to view the app's onboarding flow after
pressing "Get Started", which is what this PR was attempting to fix.